+++ game-aska/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Fixed: Bump BepInEx BE version to 755 and fix URL
 - Added: .NET 6 installation check and notification for MelonLoader
+- Added: Automatic detection of BepInEx patchers
 
 ## [0.1.0] - 2025-11-22